feature: change searchUsers to get
