+++ src/handlers/get.ts
@@ -2,7 +2,7 @@
 
 import { APIGatewayProxyHandler, APIGatewayProxyResult } from "aws-lambda";
 import { isDefined } from "../common/support";
-import { findUsers, getUser } from "../clients/dynamo-users.client";
+import { findUsers, getUser, getUsers } from "../clients/dynamo-users.client";
 import { getBingoCardFromTable } from "../clients/dynamo-bingo-card.client";
 import { getAnswers } from "../clients/dynamo-bingo-answers.client";
 
@@ -13,6 +13,24 @@ export const getCurrentUser: APIGatewayProxyHandler = async (event) => {
     response = success;
   });
   return response;
+};
+
+// POST
+// BODY PARAMS:
+// userIDs REQUIRED
+export const getFriends: APIGatewayProxyHandler = async (event) => {
+  const usernames = JSON.parse(event.queryStringParameters.usernames);
+  if (!isDefined(usernames) || usernames === []) {
+    return { statusCode: 500, body: "No body param: usernames" };
+  }
+  await getUsers(usernames).then(
+    (success) => {
+      return success;
+    },
+    (error) => {
+      return error;
+    }
+  );
 };
 
 export const searchUsers: APIGatewayProxyHandler = async (event) => {
